Fixed a few bugs concerning how we send messages out.

+++ localsms/__init__.py
@@ -4,19 +4,19 @@ import datetime
 
 import ConfigParser
 import simplejson
-
 import pygsm 
 from sqlobject import connectionForURI, sqlhub  
 import httplib2
 
-from localsms.models import Message, Log
+from localsms.db import Message, Log
 
 
-def initdb(config): 
+def initdb(config,destructive=True):
     dbfile = os.path.abspath(
         config.get("app","db_params"))
     if os.path.exists(dbfile): 
-        os.remove(dbfile)    
+        if destructive: 
+            os.remove(dbfile) 
     conn = connectionForURI(
         "%s:%s" % (config.get("app","db_type"),
                    dbfile))
@@ -41,34 +41,54 @@ def get_config(path):
     config.read(path) 
     return config 
 
-def send_sms_out(config,msg):    
+def make_http(config): 
     h = httplib2.Http() 
     h.add_credentials(
         config.get("gateway","username"),
         config.get("gateway","password"))
-    resp, content = h.request(
-        "http://%s/sms/send/" % config.get("gateway","host"),
+    return h 
+
+def send_sms_out(config,msg):
+    """
+    Makes an http request to send the message to a webserver
+    """
+    http = make_http(config)
+    resp, content = http.request(
+        "http://%s:%s/sms/send/" % ( 
+          config.get("gateway","host"),
+          config.get("gateway","port"),
+          ) ,
         "POST", body=simplejson.dumps(msg.toJson()), 
         headers={'content-type':'text/json'} )
     print resp
 
 def send_messages(config,modem): 
-    pass 
+    """
+    Send messages to the modem, checks the database for unprocesses messages
+    """
 
 def get_messages(config,modem): 
+    """
+    Takes messages off of the modem
+    """
     gsmMsg = modem.next_message()
     if gsmMsg:
         msg = Message(
+            sent=False,
+            time=datetime.datetime.now(),
             text=gsmMsg.text,
             origin=int(gsmMsg.sender))
         send_sms_out(config,msg)
     
 
 def main(*args): 
+    print("booting message pool system") 
     config = get_config("localsms.ini") 
     initdb(config) 
     modem = get_modem(config) 
     while True: 
+        print("running message pool system %s" % datetime.datetime.now()) 
         get_messages(config,modem)
         send_messages(config,modem) 
-        time.sleep(3) # put to sleep the poll 
+        time.sleep(
+            float(config.get("app","modem_poll"))) # put to sleep the poll
